US116429 - Modify how hiding indirect alignments are handled
Instead of still rendering indirect alignments and hiding them using CSS, filter them out of the array that is used by the dom-repeat.
By filtering them out, it also makes the `empty`/`isEmptyList` actually reflect what is visible. The `empty` will send an event that can be listened to by `activities` or whatever component is using this to do stuff.